Team-Management-main/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
h11==0.14.0
Flask==2.3.2
fastapi==0.103.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
uvicorn==0.23.2
idna==3.4
pydantic==2.3.0
pydantic_core==2.6.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
Werkzeug==2.3.6
python-dotenv
sqlalchemy
mysql-connector-python==8.0.28
pyodbc==5.0.1